+++ __tests__/__integration__/imperative/src/imperative.ts
@@ -66,7 +66,6 @@ export const config: IImperativeConfig = {
     defaultHome: "~/.imperative-test-cli",
     productDisplayName: "Imperative Package Test CLI",
     name: "imperative-test-cli",
-    credentialServiceName: "imperative-test-cli",
     envVariablePrefix: "IMPERATIVE_TEST_CLI",
     allowPlugins: false,
     configAutoInitCommandConfig: {