+++ packages/cmd/__tests__/CommandProcessor.test.ts
@@ -246,7 +246,7 @@ describe("Command Processor", () => {
     afterEach(() => {
         process.stdout.write = ORIGINAL_STDOUT_WRITE;
         process.stderr.write = ORIGINAL_STDERR_WRITE;
-        if (ImperativeConfig?.instance?.config?.opts?.noLoad) { delete ImperativeConfig.instance.config.opts.noLoad; }
+        jest.restoreAllMocks();
     });
 
     it("should allow us to create an instance", () => {
@@ -1398,49 +1398,10 @@ describe("Command Processor", () => {
             };
         });
 
-        const parms: any = {
-            arguments: {
-                _: ["check", "for", "banana"],
-                $0: "",
-                valid: true,
-                dcd: process.cwd()
-            },
-            silent: true
-        };
-        const commandResponse: ICommandResponse = await processor.invoke(parms);
-        expect(commandResponse).toBeDefined();
-        expect(commandResponse).toMatchSnapshot();
-    });
-
-    it("should reset noLoad on config load", async () => {
-        // Allocate the command processor
-        ImperativeConfig.instance.config.opts.noLoad = true;
-        expect(ImperativeConfig.instance.config.opts.noLoad).toEqual(true);
-
-        const processor: CommandProcessor = new CommandProcessor({
-            envVariablePrefix: ENV_VAR_PREFIX,
-            fullDefinition: SAMPLE_COMPLEX_COMMAND,
-            definition: SAMPLE_COMMAND_REAL_HANDLER,
-            helpGenerator: FAKE_HELP_GENERATOR,
-            profileManagerFactory: FAKE_PROFILE_MANAGER_FACTORY,
-            rootCommandName: SAMPLE_ROOT_COMMAND,
-            commandLine: "",
-            promptPhrase: "dummydummy"
-        });
-
-        // Mock read stdin
-        (SharedOptions.readStdinIfRequested as any) = jest.fn((args, response, type) => {
-            // Nothing to do
-        });
-
-        // Mock the profile loader
-        (CommandProfileLoader.loader as any) = jest.fn((args) => {
-            return {
-                loadProfiles: (profArgs: any) => {
-                    // Nothing to do
-                }
-            };
-        });
+        const mockConfigReload = jest.fn();
+        jest.spyOn(ImperativeConfig, "instance", "get").mockReturnValue({
+            config: { reload: mockConfigReload }
+        } as any);
 
         const parms: any = {
             arguments: {
@@ -1454,7 +1415,7 @@ describe("Command Processor", () => {
         const commandResponse: ICommandResponse = await processor.invoke(parms);
         expect(commandResponse).toBeDefined();
         expect(commandResponse).toMatchSnapshot();
-        expect(ImperativeConfig.instance.config.opts.noLoad).not.toBeDefined();
+        expect(mockConfigReload).toHaveBeenCalledTimes(1);
     });
 
     it("should extract arguments not specified on invoke from a profile and merge with args", async () => {